make z scale more than only 0-1, if not the cube is seeen incorreclty

+++ src/screenspace/screen/screen.rs
@@ -52,9 +52,9 @@ impl Screen {
         format!("{}{}{}", color.ansi_code(), "█", reset)
     }
     pub fn project_point(&self, value: &Pos3) -> ScreenPosition {
-        let x = value.x() as f64 / self.width as f64;
-        let y = value.y() as f64 / self.height as f64;
-        let z = value.z() as f64;
+        let x = value.x() / self.width as f64;
+        let y = value.y() / self.height as f64;
+        let z = value.z();
 
         if z == 0.0 {
             return ScreenPosition::with_pos(&0, &0);
